Improve Secrecy
Most of the time, you are using the bot on public servers. This causes
a problem since the mcspambot usually screams out what it is doing to
the whole server. This patch fix makes the bot reply to the owner that
is messaging it, improving the secrecy but not causing much harm